requirements.txt
k8s
jinja2
python-varnish==0.2.1
